Fix Scoring

+++ src/main/java/com/ramenreviewers/blog/model/Dish.java
@@ -9,18 +9,18 @@ import org.pkl.config.java.mapper.Named;
 public final class Dish {
 
   private  String name;
-  private  int price;
-  private  int scoreBroth;
-  private  int scoreNoodles;
-  private  int scoreToppings;
+  private  float price;
+  private  float scoreBroth;
+  private  float scoreNoodles;
+  private  float scoreToppings;
   private  String description;
 
   public Dish(
       @Named("name") String name,
-      @Named("price") int price,
-      @Named("scoreBroth") int scoreBroth,
-      @Named("scoreNoodles") int scoreNoodles,
-      @Named("scoreToppings") int scoreToppings,
+      @Named("price") float price,
+      @Named("scoreBroth") float scoreBroth,
+      @Named("scoreNoodles") float scoreNoodles,
+      @Named("scoreToppings") float scoreToppings,
       @Named("description") String description) {
     this.name = name;
     this.price = price;
